Adding `does_not_raise` to documentation only
